Mandible_Reconstruction/requirements.txt
absl-py==1.3.0
anyio==3.4.0
argon2-cffi==21.2.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.1.3
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distro==1.4.0
docker-pycreds==0.4.0
dominate==2.7.0
einops==0.7.0
entrypoints==0.3
flatbuffers==22.10.26
fonttools==4.28.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.3
imageio==2.13.3
importlib-metadata==5.0.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-tabnine==0.0.24
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
natsort==8.2.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.1
notebook==6.4.6
numexpr==2.8.0
numpy==1.21.4
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycparser==2.21
Pygments==2.10.0
PyGObject==3.36.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.3.2
setuptools-scm==6.3.2
shortuuid==1.0.9
SimpleITK==2.1.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
ssh-import-id==5.10
tables==3.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchfile==0.1.0
torchmetrics==0.11.1
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
ujson==5.4.0
urllib3==1.26.11
visdom==0.1.8.9
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.6.0
